backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
ccxt==4.4.45
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
sqlalchemy==2.0.37
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
apscheduler==3.10.4
python-multipart==0.0.20
